Python/Mega Projects/Multilingual AI Assistant/requirements.txt
SpeechRecognition
pyaudio
google-generativeai
gTTS
pipwin
streamlit

